+++ tslint.json
@@ -9,7 +9,8 @@
         "object-literal-sort-keys": false,
         "trailing-comma": [true, {"multiline": "always"}],
         "interface-name": false,
-        "no-shadowed-variable": false
+        "no-shadowed-variable": false,
+        "max-classes-per-file": false
     },
     "rulesDirectory": []
 }
